+++ benchmarks/setup.rb
@@ -1,3 +1,2 @@
-# frozen_string_literal: true
-
-require "bigdecimal"
+$LOAD_PATH.unshift(File.expand_path("../../lib", __FILE__))
+require "slow_operations"